+++ AST/Call.h
@@ -61,7 +61,7 @@ public:
 	ParamIterator begin() const { return args_.begin(); }
 	ParamIterator end() const { return args_.end(); }
 
-	virtual void PrettyPrint(Bytestream&, size_t indent = 0) const override;
+	virtual void PrettyPrint(Bytestream&, unsigned int indent = 0) const override;
 	virtual void Accept(Visitor&) const override;
 
 	virtual dag::ValuePtr evaluate(EvalContext&) const override;
