+++ AST/Foreach.h
@@ -58,7 +58,7 @@ public:
 	const Parameter& loopParameter() const { return mapping_->target(); }
 	const Expression& loopBody() const { return *body_; }
 
-	virtual void PrettyPrint(Bytestream&, size_t indent = 0) const override;
+	virtual void PrettyPrint(Bytestream&, unsigned int indent = 0) const override;
 	virtual void Accept(Visitor&) const override;
 
 	virtual dag::ValuePtr evaluate(EvalContext&) const override;
