+++ AST/Import.h
@@ -67,7 +67,7 @@ public:
 	const UniqPtrVec<Argument>& arguments() const { return arguments_; }
 	const std::string& subdirectory() const { return subdirectory_; }
 
-	virtual void PrettyPrint(Bytestream&, size_t indent = 0) const override;
+	virtual void PrettyPrint(Bytestream&, unsigned int indent = 0) const override;
 	virtual void Accept(Visitor&) const override;
 
 	virtual dag::ValuePtr evaluate(EvalContext&) const override;
