+++ Support/Printable.h
@@ -54,7 +54,7 @@ public:
 	 * Print a human-readable representation to an output stream
 	 * and return that output stream.
 	 */
-	virtual void PrettyPrint(Bytestream&, size_t indent = 0) const = 0;
+	virtual void PrettyPrint(Bytestream&, unsigned int indent = 0) const = 0;
 
 	virtual std::string str() const;
